refactor(data): move RandomCrop to new augmentation module
